+++ lib/src/risk_system/bid_ask.dart
@@ -1,0 +1,16 @@
+part of '../../risk_system.dart';
+
+enum BidAsk {
+  bid,
+  ask;
+
+  BidAsk parse(String x) {
+    if (x.toLowerCase() == 'bid') {
+      return bid;
+    } else if (x.toLowerCase() == 'ask') {
+      return ask;
+    } else {
+      throw ArgumentError('Unsupported input $x');
+    }
+  }
+}