+++ test/financial/trade_test.dart
@@ -1,7 +1,5 @@
 library test.financial.trade_test;
 
-import 'dart:ffi';
-
 import 'package:date/date.dart';
 import 'package:elec/risk_system.dart' hide Trade;
 import 'package:elec/src/financial/trading_strategy/portfolio.dart';